+++ Radish.Service/ExperienceService.cs
@@ -388,182 +388,7 @@ namespace Radish.Service;
 
     #endregion
 
-    #region 排行榜
-
-    /// <summary>
-    /// 获取等级排行榜
-    /// </summary>
-    public async Task<PageModel<LeaderboardItemVo>> GetLeaderboardAsync(
-        int pageIndex,
-        int pageSize = 50,
-        long? currentUserId = null)
-    {
-        try
-        {
-            // 限制每页数量
-            if (pageSize > 100) pageSize = 100;
-            if (pageSize < 1) pageSize = 50;
-            if (pageIndex < 1) pageIndex = 1;
-            var now = DateTime.Now;
-
-            // 查询排行榜（按 TotalExp 降序，CurrentLevel 降序）
-            var (pagedData, totalCount) = await _userExpRepository.QueryPageAsync(
-                whereExpression: e =>
-                    (!e.ExpFrozen || (e.FrozenUntil != null && e.FrozenUntil <= now)) &&
-                    e.UserId > 0 &&
-                    !e.IsDeleted,
-                pageIndex: pageIndex,
-                pageSize: pageSize,
-                orderByExpression: e => e.TotalExp,
-                orderByType: OrderByType.Desc
-            );
-
-            // 获取用户信息
-            var userIds = pagedData.Select(e => e.UserId).ToList();
-            var users = await _userRepository.QueryAsync(u => userIds.Contains(u.Id));
-            var userDict = users.ToDictionary(u => u.Id);
-
-            // 获取等级配置
-            var levels = pagedData.Select(e => e.CurrentLevel).Distinct().ToList();
-            var levelConfigs = await _levelConfigRepository.QueryAsync(l => levels.Contains(l.Level));
-            var levelConfigDict = levelConfigs.ToDictionary(l => l.Level);
-
-            // 计算起始排名
-            var startRank = (pageIndex - 1) * pageSize + 1;
-
-            // 映射为 LeaderboardItemVo
-            var leaderboard = new List<LeaderboardItemVo>();
-            var rank = startRank;
-            for (int i = 0; i < pagedData.Count; i++)
-            {
-                var exp = pagedData[i];
-                if (!userDict.TryGetValue(exp.UserId, out var user))
-                {
-                    Log.Warning("排行榜跳过不存在的用户：userId={UserId}", exp.UserId);
-                    continue;
-                }
-
-                var item = new LeaderboardItemVo
-                {
-                    VoRank = rank,
-                    VoUserId = exp.UserId,
-                    VoUserName = user.UserName,
-                    VoCurrentLevel = exp.CurrentLevel,
-                    VoCurrentLevelName = levelConfigDict.ContainsKey(exp.CurrentLevel)
-                        ? levelConfigDict[exp.CurrentLevel].LevelName
-                        : $"Lv.{exp.CurrentLevel}",
-                    VoThemeColor = levelConfigDict.ContainsKey(exp.CurrentLevel)
-                        ? levelConfigDict[exp.CurrentLevel].ThemeColor
-                        : "#9E9E9E",
-                    VoTotalExp = exp.TotalExp,
-                    VoIsCurrentUser = currentUserId.HasValue && exp.UserId == currentUserId.Value
-                };
-
-                leaderboard.Add(item);
-                rank++;
-            }
-
-            var pageCount = (int)Math.Ceiling(totalCount / (double)pageSize);
-
-            return new PageModel<LeaderboardItemVo>
-            {
-                Page = pageIndex,
-                PageSize = pageSize,
-                DataCount = totalCount,
-                PageCount = pageCount,
-                Data = leaderboard
-            };
-        }
-        catch (Exception ex)
-        {
-            Log.Error(ex, "获取排行榜失败");
-            throw;
-        }
-    }
-
-    /// <summary>
-    /// 获取用户排名
-    /// </summary>
-    public async Task<int> GetUserRankAsync(long userId)
-    {
-        try
-        {
-            if (userId <= 0)
-            {
-                return 0;
-            }
-
-            var userExp = await _userExpRepository.QueryFirstAsync(e => e.UserId == userId && !e.IsDeleted);
-            if (userExp == null)
-            {
-                return 0;
-            }
-
-            userExp = await NormalizeFreezeStateAsync(userExp);
-            if (IsFreezeActive(userExp, DateTime.Now))
-            {
-                return 0; // 未上榜或仍处于冻结中
-            }
-
-            var now = DateTime.Now;
-
-            // 统计比该用户经验值高的用户数量
-            var higherCount = await _userExpRepository.QueryCountAsync(
-                e =>
-                    (!e.ExpFrozen || (e.FrozenUntil != null && e.FrozenUntil <= now)) &&
-                    e.UserId > 0 &&
-                    e.TotalExp > userExp.TotalExp &&
-                    !e.IsDeleted
-            );
-
-            return (int)higherCount + 1; // 排名 = 比自己高的数量 + 1
-        }
-        catch (Exception ex)
-        {
-            Log.Error(ex, "获取用户 {UserId} 排名失败", userId);
-            return 0;
-        }
-    }
-
-    #endregion
-
     #region 管理员操作
-
-    /// <summary>
-    /// 管理员调整用户经验值
-    /// </summary>
-    public async Task<bool> AdminAdjustExperienceAsync(
-        long userId,
-        int deltaExp,
-        string reason,
-        long operatorId,
-        string operatorName)
-    {
-        if (userId <= 0)
-        {
-            Log.Warning("管理员调整经验失败：userId 无效（{UserId}）", userId);
-            return false;
-        }
-
-        if (deltaExp == 0)
-        {
-            Log.Warning("管理员调整经验失败：变动量不能为 0，userId={UserId}", userId);
-            return false;
-        }
-
-        var normalizedReason = string.IsNullOrWhiteSpace(reason) ? "管理员调整" : reason.Trim();
-
-        try
-        {
-            return await ExecuteWithRetryAsync(async () =>
-                await AdminAdjustExperienceInternalAsync(userId, deltaExp, normalizedReason, operatorId, operatorName));
-        }
-        catch (Exception ex)
-        {
-            Log.Error(ex, "管理员调整用户 {UserId} 经验失败，deltaExp={DeltaExp}", userId, deltaExp);
-            return false;
-        }
-    }
 
     /// <summary>
     /// 记录人工复核结论
@@ -773,119 +598,6 @@ namespace Radish.Service;
             Log.Error(ex, "用户 {UserId} 经验值记录初始化失败且无法查询到已存在记录", userId);
             return null;
         }
-    }
-
-    [UseTran(Propagation = Propagation.Required)]
-    private async Task<bool> AdminAdjustExperienceInternalAsync(
-        long userId,
-        int deltaExp,
-        string reason,
-        long operatorId,
-        string operatorName)
-    {
-        var userExp = await _userExpRepository.QueryFirstAsync(e => e.UserId == userId && !e.IsDeleted);
-        if (userExp == null)
-        {
-            userExp = await InitializeUserExperienceAsync(userId);
-            if (userExp == null)
-            {
-                Log.Error("管理员调整经验失败：用户 {UserId} 经验值记录初始化失败", userId);
-                return false;
-            }
-        }
-
-        userExp = await NormalizeFreezeStateAsync(userExp);
-
-        var levelConfigs = await GetLevelConfigsCacheAsync();
-        var oldTotalExp = userExp.TotalExp;
-        var oldLevel = userExp.CurrentLevel;
-        var newTotalExp = oldTotalExp + deltaExp;
-        var actualDelta = deltaExp;
-
-        if (newTotalExp < 0)
-        {
-            actualDelta = -(int)oldTotalExp;
-            newTotalExp = 0;
-        }
-
-        if (actualDelta == 0)
-        {
-            Log.Warning("管理员调整经验未生效：userId={UserId}, oldTotalExp={OldTotalExp}, deltaExp={DeltaExp}",
-                userId, oldTotalExp, deltaExp);
-            return false;
-        }
-
-        var (newLevel, newCurrentExp) = CalculateLevel(newTotalExp, levelConfigs);
-        var now = DateTime.Now;
-
-        var updatedRows = await _userExpRepository.UpdateColumnsAsync(
-            e => new UserExperience
-            {
-                CurrentLevel = newLevel,
-                CurrentExp = newCurrentExp,
-                TotalExp = newTotalExp,
-                LevelUpAt = newLevel > oldLevel ? now : e.LevelUpAt,
-                Version = e.Version + 1,
-                ModifyTime = now,
-                ModifyBy = operatorName,
-                ModifyId = operatorId
-            },
-            e => e.UserId == userId && e.Version == userExp.Version && !e.IsDeleted
-        );
-
-        if (updatedRows == 0)
-        {
-            throw new ConcurrencyException("乐观锁冲突：管理员调整经验时记录已被其他操作修改");
-        }
-
-        var transaction = new ExpTransaction
-        {
-            Id = SnowFlakeSingle.instance.getID(),
-            UserId = userId,
-            ExpType = actualDelta > 0 ? "ADMIN_ADJUST" : "PENALTY",
-            ExpAmount = actualDelta,
-            BusinessType = "User",
-            BusinessId = userId,
-            Remark = reason,
-            ExpBefore = oldTotalExp,
-            ExpAfter = newTotalExp,
-            LevelBefore = oldLevel,
-            LevelAfter = newLevel,
-            CreatedDate = DateTime.Today,
-            TenantId = userExp.TenantId,
-            CreateTime = now,
-            CreateBy = operatorName,
-            CreateId = operatorId
-        };
-        await _expTransactionRepository.AddAsync(transaction);
-
-        Log.Information(
-            "管理员 {OperatorName}({OperatorId}) 调整用户 {UserId} 经验成功，实际变动 {ActualDelta}，总经验 {OldTotalExp} -> {NewTotalExp}",
-            operatorName,
-            operatorId,
-            userId,
-            actualDelta,
-            oldTotalExp,
-            newTotalExp);
-
-        if (newLevel > oldLevel)
-        {
-            Log.Information("管理员调整触发升级：用户 {UserId} 从 Lv.{OldLevel} 升级到 Lv.{NewLevel}", userId, oldLevel, newLevel);
-            _ = Task.Run(async () =>
-            {
-                try
-                {
-                    await HandleLevelUpAsync(userId, oldLevel, newLevel);
-                }
-                catch (Exception ex)
-                {
-                    Log.Error(ex, "处理管理员调整后的升级事件失败：userId={UserId}, oldLevel={OldLevel}, newLevel={NewLevel}",
-                        userId, oldLevel, newLevel);
-                }
-            });
-        }
-
-        return true;
     }
 
     /// <summary>
